fix: bean Int 溢出导致登录报错与发现页闪退
- LoginStatusData.Account.id: Int -> Long（11 位新用户 ID 超出 Int 范围，登录报 Failed to parse int）
- SongData.publishTime/mark/djId/sId/mv: Int -> Long（13 位毫秒时间戳与 11 位 mark 导致每日推荐/搜索/歌单列表解析崩溃）
- OriginSongSimpleData.songId: Int -> Long
- QualityData.size: Int -> Long（无损音频字节数防御）

+++ shared/src/commonMain/kotlin/me/wcy/music/common/bean/OriginSongSimpleData.kt
@@ -10,7 +10,7 @@ import kotlinx.serialization.Transient
 @Serializable
 data class OriginSongSimpleData(
     @SerialName("songId")
-    val songId: Int = 0,
+    val songId: Long = 0,
     @SerialName("name")
     val name: String = "",
     @SerialName("artists")
